Use secure password encryption with ssh-keygen
This uses the modern secure encryption method for passwords: https://latacora.singles/2018/08/03/the-default-openssh.html

Even if the user does not use a password this time, this command is found in the top 20 Google search results for creating ssh keys. It would be extremely beneficial to others to get into the habit of using the `-o` option every time they create new keys, even if they aren't setting a password.